pyproject.toml
[tool.poetry]
name = "delivery-service"
version = "0.1.0"
description = ""
authors = ["Nikita <frostandearth@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0.37"
aiomysql = "^0.2.0"
aioredis = "^2.0.1"
pydantic = "^2.10.6"
httpx = "^0.28.1"
apscheduler = "^3.11.0"
python-dotenv = "^1.0.1"
redis = "^5.2.1"
jinja2 = "^3.1.5"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
